Implement gRPC customizations another way without calling reinterpret_cast on the client and server C++ types
